Added some javadoc in the WirelessConfiguration class.
Now, we have to set the channel name when we create it, in order to
prevent the nullexceptions that could happen in a situation where we
don't have the channel name.
The method createChannel needs only the channel and no longer his name
(cause it's already saved in the channel object)
Started adding a new argument to the command /WRI, which will allow to
see the position of each channel and its status.