I need to make the debug mode less... verbose
No seriously, that line could put a server offline by crashing, because
it was puttin a lot of messages in the debug mode. Imagine, there are a
lot of chunks which are prevented from unloading for ONE sign, so on a
big server... that creates a lot of messages, and the logs become messy.
I have able to see it on Kervinou's server.