Rework of onBlockRedstoneChange method
This rework of the onBlockRedstoneChange method uses the scheduler to schedule recursive calls until the sign or no sign is found. If the sign is found, the channel is handled the same way as before. The reason for using the scheduler is to provide time for the sign blocks isBlockIndirectlyPowered and isBlockPowered methods to update and be useful as it appears the event is usually called before those values are updated and, therefore, blocks their use.